Corrigiendo Bugs

+++ src/controllers/productController.js
@@ -9,7 +9,7 @@ export const listarProductos = async (req, res) => {
     // const productos = await Producto.find({ user: req.user.id }).populate("User");
     res.json(productos);
   } catch (error) {
-    res.status(500).json({ error: 'Error al obtener la lista de productos' });
+    res.status(400).json({ error: 'Error al obtener la lista de productos' });
   }
 };
 
@@ -19,44 +19,9 @@ export const obtenerProducto = async (req, res) => {
     if (!producto) return res.status(404).json({ message: 'Producto no encontrado' });
     res.json(producto);
   } catch (error) {
-    res.status(500).json({ error: 'Error al obtener el producto' });
+    res.status(400).json({ error: 'Error al obtener el producto' });
   }
 };
-
-// export const crearProducto = async (req, res) => {
-//   try {
-//     const { nombre, descripcion, categoria, subcategoria } = req.body;
-//     const precio = parseFloat(req.body.precio);
-//     const isDisponible = req.body.disponible === 'true';
-//     const ingredientes = req.body.ingredientes.split(',').map((ingrediente) => ingrediente.trim());
-//     const campoProducto = {
-//       nombre,
-//       descripcion,
-//       precio,
-//       ingredientes,
-//       categoria,
-//       subcategoria,
-//       disponible: isDisponible,
-//     };
-//     if (req.files?.image) {
-//       const result = await subirImagen(req.files.image.tempFilePath);
-//       campoProducto.imagen = {
-//         public_id: result.public_id,
-//         secure_url: result.secure_url,
-//       };
-//       await fs.unlink(req.files.image.tempFilePath);
-//       // console.log(result);
-//     }
-//     const validData = productoSchema.parse(campoProducto);
-//     const producto = new Producto({ ...validData, user: req.user.id });
-//     const productoCreado = await producto.save();
-//     console.log(productoCreado);
-//     res.json(productoCreado);
-//   } catch (error) {
-//     console.log(error);
-//     return res.status(400).json(error.errors.map((error) => error.message));
-//   }
-// };
 
 export const crearProducto = async (req, res) => {
   try {
@@ -74,46 +39,49 @@ export const crearProducto = async (req, res) => {
       disponible: isDisponible,
     };
 
-    const validData = productoSchema.parse(campoProducto);
+    const validarCampos = productoSchema.parse(campoProducto);
+    const producto = new Producto({ ...validarCampos, user: req.user.id });
+
+    const productoExiste = await Producto.findOne({ nombre: validarCampos.nombre });
+    if (productoExiste)
+      return res.status(409).json({ message: 'El nombre del producto ya existe. Por favor, elija otro nombre.' });
+
     if (req.files?.image) {
       const result = await subirImagen(req.files.image.tempFilePath);
-      campoProducto.imagen = {
+      producto.imagen = {
         public_id: result.public_id,
         secure_url: result.secure_url,
       };
       await fs.unlink(req.files.image.tempFilePath);
     }
-    const producto = new Producto({ ...validData, user: req.user.id });
+
     const productoCreado = await producto.save();
     console.log(productoCreado);
     res.json(productoCreado);
   } catch (error) {
     console.log(error);
     return res.status(400).json(error.errors.map((error) => error.message));
+  } finally {
+    if (req.files?.image) {
+      await fs.unlink(req.files.image.tempFilePath).catch((error) => {
+        console.log("Error al eliminar la imagen temporal:", error);
+      });
+    }
   }
 };
 
 
 export const actualizarProducto = async (req, res) => {
   try {
-    const {
-      nombre,
-      descripcion,
-      imagen,
-      categoria,
-      subcategoria,
-      disponible,
-      precio,
-      ingredientes,
-    } = req.body;
+    const { nombre, descripcion, imagen, categoria, subcategoria, disponible, precio, ingredientes } = req.body;
     const isDisponible = req.body.disponible === 'true';
+
     const actualizarCampo = {};
     if (nombre) actualizarCampo.nombre = nombre;
     if (descripcion) actualizarCampo.descripcion = descripcion;
     if (imagen) actualizarCampo.imagen = imagen;
     if (categoria) actualizarCampo.categoria = categoria;
     if (subcategoria) actualizarCampo.subcategoria = subcategoria;
-    // if (disponible !== undefined) actualizarCampo.disponible = Boolean(disponible);
     if (disponible) actualizarCampo.disponible = isDisponible;
     if (precio !== undefined) actualizarCampo.precio = parseFloat(precio);
     if (ingredientes) {
@@ -121,32 +89,47 @@ export const actualizarProducto = async (req, res) => {
         .split(',')
         .map((ingrediente) => ingrediente.trim());
     }
-    const validData = productoSchema.parse(actualizarCampo);
+
+    const validarCampos = productoSchema.parse(actualizarCampo);
     const producto = await Producto.findById(req.params.id);
+
+    const productoExiste = await Producto.findOne({ nombre: validarCampos.nombre });
+    if (productoExiste && productoExiste._id.toString() !== producto._id.toString()) {
+      return res.status(409).json({ message: 'El nombre del producto ya existe. Por favor, elija otro nombre.' });
+    }
+
     if (req.files?.image) {
-      await eliminarImagen(producto.imagen.public_id);
-      const result = await actualizarImagen(
-        producto.imagen.public_id,
-        req.files.image.tempFilePath
-      );
-      producto.imagen = {
-        public_id: result.public_id,
-        secure_url: result.secure_url,
-      };
+      if (producto.imagen && producto.imagen.public_id) {
+        await eliminarImagen(producto.imagen.public_id);
+        const result = await actualizarImagen(
+          producto.imagen.public_id,
+          req.files.image.tempFilePath
+        );
+        producto.imagen = {
+          public_id: result.public_id,
+          secure_url: result.secure_url,
+        };
+      }
       await fs.unlink(req.files.image.tempFilePath);
     }
+
     const productoId = req.params.id;
     const productoActualizado = await Producto.findOneAndUpdate(
       { _id: productoId, user: req.user.id },
-      validData,
+      validarCampos,
       { new: true }
     );
     if (!productoActualizado) return res.status(404).json({ error: 'Producto no encontrado' });
-
     res.json(productoActualizado);
   } catch (error) {
     console.log(error);
     return res.status(400).json(error.errors.map((error) => error.message));
+  } finally {
+    if (req.files?.image) {
+      await fs.unlink(req.files.image.tempFilePath).catch((error) => {
+        console.log("Error al eliminar la imagen temporal:", error);
+      });
+    }
   }
 };
 
@@ -154,9 +137,11 @@ export const eliminarProducto = async (req, res) => {
   try {
     const producto = await Producto.findByIdAndDelete(req.params.id);
     if (!producto) return res.status(404).json({ message: 'Producto no encontrado' });
-    await eliminarImagen(producto.imagen.public_id);
+    if (producto.imagen && producto.imagen.public_id)
+      await eliminarImagen(producto.imagen.public_id);
+
     res.json(producto);
   } catch (error) {
-    res.status(500).json({ error: 'Error al eliminar el producto' });
+    res.status(400).json({ error: 'Error al eliminar el producto' });
   }
 };
